+++ CHANGES.md
@@ -10,6 +10,7 @@ Release Notes.
 - Add RBAC files and service account to support Kubernetes coordination.
 - Add default and validation webhooks to operator controllers.
 - Add UI CRD to deploy skywalking UI server.
+- Add Fetcher CRD to fetch metrics from other telemetry system, for example, Prometheus.
 
 #### Chores
 - Transform project layers to support multiple applications.